fix: weird format when ssh into other machines from ghostty
After trying to ssh into some machines at work, I observed that the
shell became unusable (i.e., a lot of duplicate words and garbage
being printed). The issue, as explained in [1], seems to be that
ghostty sets the environment variable TERM to xterm-ghostty, which
the remote hosts have no idea how to handle.

To solve the above, and again as explained in [1], I set the variable
TERM to xterm-256color if I am using ghostty.

Alternative solutions would be to use SetEnv TERM=xterm-256color in
each entry of .ssh/config, but then things might break if a try to
ssh into a host not listed there.

[1] https://vninja.net/2024/12/28/ghostty-workaround-for-missing-or-unsuitable-terminal-xterm-ghostty/