fix: export correct bin folder regardless of username
In Linux I use pipx to install different packages, such as Conan.
This tool install binaries by default in /home/<user>/.local/bin,
and this needs to be added to the path. Since I want to make this
independent of my username, now I set this with $USER.

I also moved the entry to the "Linux-only" section, as this is not
relevant for MacOS.