fix: automatically configure ghostty for ssh
It might be that the Terminfo populated by ghostty is not supported
in machines with old ncurses packages. The solution, as described
in [1], is to either copy Ghostty's terminfo entry to the remote
machine, or configure SSH to fall back to a known terminfo. We
can now do this automatically by adding the specific option.

[1] https://ghostty.org/docs/help/terminfo